--- plugin.yml
@@ -1,4 +1,0 @@
-name: StopBucketFill
-main: me.noodles.sbf.MainSBF
-version: 3.0
-
