+++ src/main/java/me/noodles/sbf/UpdateJoinEvent.java
@@ -14,14 +14,14 @@ public class UpdateJoinEvent implements Listener
     public void onJoin(final PlayerJoinEvent e) {
     	Player p = e.getPlayer();
     	if (p.hasPermission("sbf.update")) {
-    		if (MainSBF.getPlugin().getConfig().getBoolean("Update.Enabled") == true){
-    		this.checker = new UpdateChecker(MainSBF.plugin);
+    		if (SBF.getPlugin().getConfig().getBoolean("Update.Enabled") == true){
+    		this.checker = new UpdateChecker(SBF.plugin);
                         if (this.checker.isConnected()) {
                             if (this.checker.hasUpdate()) {
                             	p.sendMessage(ChatColor.GRAY + "=========================");
                                 p.sendMessage(ChatColor.RED + "StopBucketFill is outdated!");
                                 p.sendMessage(ChatColor.GREEN + "Newest version: " + this.checker.getLatestVersion());
-                                p.sendMessage(ChatColor.RED + "Your version: " + MainSBF.plugin.getDescription().getVersion());
+                                p.sendMessage(ChatColor.RED + "Your version: " + SBF.plugin.getDescription().getVersion());
                                 p.sendMessage(ChatColor.GRAY + "=========================");
                             }
                         }
